Move some serialization code to a sub-module
